hwcaps: Change uint32_t to unsigned long
getauxval() returns an unsigned long which can be defined to be
more than 32 bits so don't force the result into a uint32_t.

+++ include/internal/hwcaps.h
@@ -115,8 +115,8 @@ CPU_FEATURES_START_CPP_NAMESPACE
 
 
 typedef struct {
-  uint32_t hwcaps;
-  uint32_t hwcaps2;
+  unsigned long hwcaps;
+  unsigned long hwcaps2;
 } HardwareCapabilities;
 
 HardwareCapabilities GetHardwareCapabilities(void);
